new pro/requirements.txt
Flask==2.1.2
opencv-python==4.5.5.62
pytesseract==0.3.8
Pillow==8.4.0
